requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.11
packaging==26.0
pillow==12.2.0
python-dotenv==1.0.1
requests==2.33.1
typing_extensions==4.12.2
urllib3==2.6.3
